Update: completely using bwesterb structure

+++ params.go
@@ -5,7 +5,7 @@ package mbpqs
 */
 type Params struct {
 	n         uint32 // the security parameter, length of message digest and three nodes in bytes.
-	w         uint32 // the Winternitz parameter, used in WOTS-T.
+	w         uint16 // the Winternitz parameter, used in WOTS-T.
 	rootH     uint32 // the height of the three (# levels -1).
 	initChanH uint32 // the inital chain tree height.
 	d         uint32 // the chain tree height growing exponent.
